backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
openai==1.59.4
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.28.1
python-dotenv
python-multipart
